+++ comics/settings/base.py
@@ -161,11 +161,6 @@ CACHE_MIDDLEWARE_ANONYMOUS_ONLY = True
 DATE_FORMAT = 'D j M Y'
 TIME_FORMAT = 'H:i'
 
-# Haystack settings, only apply if comics.search and haystack are used
-HAYSTACK_SITECONF = 'comics.search.indexes'
-HAYSTACK_SEARCH_ENGINE = 'solr'
-HAYSTACK_SOLR_URL = 'http://127.0.0.1:8983/solr'
-
 # django_compressor settings
 # Explicitly use HtmlParser to avoid depending on BeautifulSoup through the use
 # of LxmlParser